+++ CHANGELOG.md
@@ -22,11 +22,18 @@
 - Added `Projection` method to `Vector2` and `Vector3`.
 
 ### Changed
+
 - `Destroy()` event triggers even if an entity is disabled.
+- Migrated to NET6.
+- Updated MonoGame to `3.8.1`.
+- Updated Nopipeline to `2.2.0`.
+- Replaced `Newtonsoft.Json` dependency with `System.Text.Json`.
+- Removed VS2019 support.
 
 ### Fixed
 
 - Fixed `Sprite.Clone()` not copying frames properly.
+
 
 
 ## [v2.1.0] - *12.08.2021*